end-to-end-solutions/Luna/src/Luna.Agent/Luna.Agent/requirements.txt
azureml-sdk
flask
requests
luna-publish-utils
sqlalchemy
pyodbc
azure-keyvault-secrets
azure.identity
azure-storage-blob